requirements.txt
returns
pytest
assertpy
python-box